+++ G12P2/src/main/java/G12P2/ag/seleccion/Seleccion.java
@@ -3,5 +3,5 @@ package G12P2.ag.seleccion;
 import G12P2.cromosomas.Cromosoma;
 
 public interface Seleccion {
-    Cromosoma[] seleccionar(Cromosoma[] poblacion, int[] fitness);
+    Cromosoma[] seleccionar(Cromosoma[] poblacion, double[] fitness);
 }
